pyproject.toml
[project]
name = "postgres-mcp"
version = "0.1.0"
description = "Multi-connection PostgreSQL MCP server for Claude Code"
authors = [
    {name = "Jake"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli] (>=1.26.0,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]

[project.scripts]
postgres-mcp = "postgres_mcp.server:main"

[[tool.poetry.packages]]
include = "postgres_mcp"
from = "src"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
